% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{loads the data from the Internet and returns a data frame}
\usage{
load_data(url)
}
\arguments{
\item{url}{the path to the input file (a string of a URL)}
}
\value{
a data frame with the data from the given URL
}
\description{
loads the data from the Internet and returns a data frame
}
\examples{
load_data("https://7e6cd356-86ad-4874-abc7-3a69bbbc39e6.filesusr.com/ugd/c5a545_c1b17c070c984dfcb14cf1c3bb0b6e67.csv?dn=garments_worker_productivity.csv")
}
